+++ lib/src/widgets/auto_hide_play_pause_button.dart
@@ -50,6 +50,8 @@ class AutoHidePlayPauseButton extends StatelessWidget {
             showReplayButton:
                 options.playerUIVisibilityOptions.showReplayButton,
             autoMuteOnStart: options.videoSourceConfiguration.autoMuteOnStart,
+            onReplay: callbacks.onReplay,
+            onFinished: callbacks.onFinished,
           ),
         );
       },